feat: update cart quantity functionality

+++ src/services/auth/signIn.js
@@ -5,9 +5,9 @@ export const signIn = async ({
     data,
     userDataDispatch,
     productDispatch,
-    setLoginState,
+    setIsUserLoggedIn,
     showToast,
-    navigate
+    keepMeLoggedIn
 }) => {
     try {
         const res = await axios.post("/api/auth/login", data);
@@ -26,15 +26,16 @@ export const signIn = async ({
                     firstName, lastName, email, createdAt, cart, wishlist, addresses
                 }
             })
-            setLoginState(true);
-            localStorage.setItem("token", encodedToken);
+            setIsUserLoggedIn(true);
             axios.defaults.headers.common["authorization"] = encodedToken;
-            localStorage.setItem('user', JSON.stringify({
-                firstName, lastName, email, createdAt, addresses
-            }))
             productDispatch({ type: "ADD_CART_WISHLIST", payload: { cart, wishlist } });
-            navigate(-1);
             showToast({ title: 'logged in successfully', type: 'success' });
+            if (keepMeLoggedIn) {
+                localStorage.setItem("token", encodedToken);
+                localStorage.setItem('user', JSON.stringify({
+                    firstName, lastName, email, createdAt, addresses
+                }))
+            }
         }
     } catch (e) {
         setSigninError(e.response.data.errors);
